% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrStrata.R
\name{aggrStrata}
\alias{aggrStrata}
\title{aggrStrata}
\usage{
aggrStrata(strata, nvar, vett, censiti, dominio)
}
\arguments{
\item{strata}{name of the dataframe containing the strata to be aggregated.}

\item{nvar}{number of target variables Y}

\item{vett}{vector of integers of the same length of the dimension of the 'strata' dataframe indicating how the strata must be aggregated.}

\item{censiti}{flag indicating if the strata are take-all (=1) or not (=0)}

\item{dominio}{variable in the strata indicating the domain level}
}
\value{
a dataframe containing the aggregated strata
}
\description{
Function to aggregate the information from a set of strata
}
\details{
Be aware that this function is applicable only to strata of a same domain level
}
\examples{
data(beat.example)
vett <- c(rep(1,5),rep(2,5),rep(3,7))
R2BEAT:::aggrStrata(strata,vett,nvar=2,dominio="DOM1",censiti=0)
}
