% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat.1cv_new.R
\name{beat.1cv_new}
\alias{beat.1cv_new}
\title{The function returns a dataframe with planned and actual coefficients of variation (CV) in a multivariate multi-domain allocation problem.}
\usage{
beat.1cv_new(stratif, alloc, minnumstrat = 2)
}
\arguments{
\item{stratif}{A dataframe with strata information.}

\item{alloc:}{A vector with an allocation into strata.}

\item{minnumstrat:}{Optional. A number indicating the lower bound for the number of allocated units in each stratum (default is 2).}
}
\value{
A dataframe with actual CV for each combination of domain, domain category and target variable.
}
\description{
The function returns a dataframe with planned and actual coefficients of variation (CV) in a multivariate multi-domain allocation problem.
}
