% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat.1st_new.R
\name{beat.1st}
\alias{beat.1st}
\title{Compute multivariate optimal allocation for different domains in one stage stratified sample design}
\usage{
beat.1st(file_strata, errors, var_label=NULL, minnumstrat = 2, maxiter = 200, maxiter1 = 25, epsilon = 1e-11)
}
\arguments{
\item{file_strata:}{Data frame of survey strata, for more details see, e.g.,\code{\link{strata}}.}

\item{errors:}{Data frame of expected coefficients of variation (CV) for each domain, for more details see, e.g.,\code{\link{errors}}.}

\item{var_label:}{a vector listing the target variable(s) which are used for leading the optimal allocation.}

\item{minnumstrat:}{Minimum number of elementary units per strata (default=2).}

\item{maxiter:}{Maximum number of iterations (default=200) of the general procedure. This kind of iteration may be required by the fact that when in a stratum the number of allocated units is greater or equal to its population, that stratum is set as "census stratum", and the whole procedure is re-initialised.}

\item{maxiter1:}{Maximum number of iterations in Chromy algorithm (default=25).}

\item{epsilon:}{Tollerance for the maximum absolute differences between the expected CV and the realised CV with the allocation obtained in the last iteraction for all domains. The default is 10^(-11).}
}
\value{
It returns a list with the following objects: a dataframe, *file_strata*, with the strata information that is one of the inputs for performing the multivariate and multidomain optimal allocation implemented in the function *beat.1st* in R2BEAT R-package; a vector, *var_label*, listing the target variable(s) which will be used for leading the optimal allocation.
#' @details
The methodology is a generalization of Bethel multivariate allocation (1989) that extends 
the Neyman (1959) - Tchuprov (1923) allocation for multi-purpose and multi-domain surveys. 
The generalized Bethel’s algorithm allows to determine the optimal sample size for each stratum 
in a stratified sample design. The overall sample size and the allocation among the different strata 
is determined starting from the accuracy constraints imposed in the survey on interest estimates.
}
\description{
This function computes the multivariate optimal allocation for different domains 
in a one-stage file_strataied sample design, following the generalized Bethel algorithm.
}
