% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat.2st_new.R
\name{beat.2st}
\alias{beat.2st}
\title{Compute multivariate optimal allocation for different domains in two-stage stratified sample design}
\usage{
beat.1st(file_strata,errors,des_file,psu_file,rho,deft_start,effst,epsilon1,mmdiff_deft,maxi,epsilon,minPSUstrat,minnumstrat,maxiter,maxiter1)
}
\arguments{
\item{file_strata:}{dataframe of survey strata, for more details see, e.g.,\code{\link{strata}}.}

\item{errors:}{dataframe of coefficients of variation for each domain, for more details see, e.g.,\code{\link {errors}}.}

\item{des_file:}{dataframe containing information on sampling design variables, for more details see, e.g.,\code{\link {design}}.}

\item{psu_file:}{dataframe containing information on primary stage units stratification, for more details see, e.g.,\code{\link {PSU_strat}}.}

\item{rho:}{dataframe of survey strata, for more details see, e.g.,\code{\link rho}.}

\item{deft_start:}{dataframe of survey strata, for taking into account the initial design effect on each variable, for more details see, e.g.,\code{\link {deft_start}}.}

\item{effst:}{dataframe of survey strata, for taking into account the estimator effect on each variable, for more details see, e.g.,\code{\link {effst}}.}

\item{epsilon1:}{first stop condition: sample sizes differences beetween two iterations; iteration continues until the maximum of sample sizes differences is greater than the default value. The default is 5.}

\item{mmdiff_deft:}{second stop condition: defts differences beetween two iterations; iteration continues until the maximum of defts largest differences is greater than the default value. The default is 0.06.}

\item{maxi:}{third stop condition: maximum number of allowed iterations. The default is 20.}

\item{epsilon:}{the same as in function \code{\link {beat.1st}}.}

\item{minPSUstrat:}{minimum number of non-self-represenative PSUs to be selected in each stratum.}

\item{minnumstrat:}{the same as in function \code{\link {beat.1st}}.}

\item{maxiter:}{the same as in function \code{\link {beat.1st}}.}

\item{maxiter1:}{the same as in function \code{\link{beat.1st}}.}
}
\value{
Returns a list with four components:

- iteractions: dataframe that for each iteraction provides a summary with the number of Primary Stage Units (\code{PSU_Total}) distinguish between Self-Representative (\code{PSU_SR}) from Non-Self-Representative (\code{PSU_NSR}) and the number of Secondary Stage Units (\code{SSU}). This output is also printed to the screen.

- file_strata: Input dataframe in \code{file_strata} with the design effect for each variables in each stratum (\code{DEFT1 - DEFTn}) and the optimal sample size columns.

- alloc: dataframe with optimal (\code{OPT}), proportional (\code{PROP}), equal (\code{UNIF}) sample size allocation.

- planned: Data frame with a summary of expected coefficients of variation for each variable in each domain.

- expected: Data frame with a summary of realized coefficients of variation with the given optimal allocation for each variable in each domain.

- sensitivity: Data frame with a summary of the sensitivity at 10\% for each domain and each variable. Sensitivity can be a useful tool to help in finding the best allocation, because it provides a hint of the expected sample size variation for a 10\% change in planned CVs.

- deft_c: Data frame with the design effect for each variable in each domain in each iteraction. Note that \code{DEFT1_0 - DEFTn_0} is always equal to 1 if \code{deft_start} is \code{NULL}. Instead is equal to \code{deft_start}. While \code{DEFT1 - DEFTn} are the final design effect related to the given allocation.

- param_alloc: A vector with a resume of all the parameter given for the allocation.
}
\description{
This function computes the multivariate optimal allocation for different domains 
in a one-stage file_strataied sample design, following the generalized Bethel algorithm.
}
\examples{

# Load example data
data(beat.example)

allocation2st_1 <- beat.2st(file_strata=strata, errors=errors,
                           des_file=design, psu_file=PSU_strat,rho=rho)

allocation2st_1$sensitivity
errors1 <- errors 
errors1[1,2] <- errors[1,2]+errors[1,2]*0.1
allocation2st_2 <- beat.2st(file_strata=strata, errors=errors1,
                           des_file=design, psu_file=PSU_strat,rho=rho)

errors2 <- errors 
errors2[1,2] <- errors[1,2]-errors[1,2]*0.1
allocation2st_3 <- beat.2st(file_strata=strata, errors=errors2,
                          des_file=design, psu_file=PSU_strat,rho=rho)
design1 <- design 
design1[,4] <- 60
allocation2st_4 <- beat.2st(file_strata=strata, errors=errors2,
                           des_file=design1, psu_file=PSU_strat, rho=rho)
allocation2st_3$expected
allocation2st_4$expected
design2 <- design 
design2[,4] <- 24
allocation2st_5 <- beat.2st(file_strata=strata, errors=errors2,
                          des_file=design2, psu_file=PSU_strat, rho=rho)
allocation2st_4$expected
allocation2st_5$expected
design
design1
design2
design3 <- design 
design3$DELTA <- 2.31
allocation2st_6 <- beat.2st(file_strata=strata, errors=errors,
                           des_file=design3, psu_file=PSU_strat, rho=rho)
library(readr)
pop <- read_rds("https://github.com/barcaroli/R2BEAT_workflows/blob/master/pop.RDS?raw=true")
library(R2BEAT)
cv <- as.data.frame(list(DOM=c("DOM1","DOM2"),
                        CV1=c(0.02,0.03),
                        CV2=c(0.03,0.06),
                        CV3=c(0.03,0.06),
                        CV4=c(0.05,0.08)))
cv
samp_frame <- pop
samp_frame$one <- 1
id_PSU <- "municipality"  
id_SSU <- "id_ind"        
strata_var <- "stratum"   
target_vars <- c("income_hh","active","inactive","unemployed")   
deff_var <- "stratum"     
domain_var <- "region"  
delta =  1       # households = survey units
minimum <- 50    # minimum number of SSUs to be interviewed in each selected PSU
deff_sugg <- 1.5 # suggestion for the deff value

inp <- prepareInputToAllocation1(samp_frame,
                               id_PSU,
                                id_SSU,
                                strata_var,
                                target_vars,
                                deff_var,
                                domain_var,
                                minimum,
                                delta,
                                deff_sugg)
inp$desfile$MINIMUM <- 50
alloc <- beat.2st(file_strata = inp$strata, 
                 errors = cv, 
                 des_file = inp$des_file, 
                 psu_file = inp$psu_file, 
                 rho = inp$rho, 
                 deft_start = NULL,
                 effst = inp$effst, 
                 minPSUstrat = 2,
                 minnumstrat = 50)
                 
}
