% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dummy_variables.R
\name{build_dummy_variables}
\alias{build_dummy_variables}
\title{build_dummy_variables}
\usage{
build_dummy_variables(frame, domain_var, initial_target_vars, cv)
}
\arguments{
\item{frame}{the sampling frame.}

\item{domain_var}{the indication of the variable that indicates the domains of interest.}

\item{initial_target_vars}{a vector containing the names of the current target variables.}

\item{cv}{the set of the current precision constraints}
}
\value{
a list containing: (a) the new frame with the values of the dummy variables; 
(b) the new target variables; (c) the new set of precision constraints
}
\description{
Function to derive dummy variables from the target variables: it is useful when it is necessary
to differentiate the precision constraints in the different domains of interest
}
