% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs_hint.R
\name{CVs_hint}
\alias{CVs_hint}
\title{cvs_hint}
\usage{
CVs_hint(strata, cv)
}
\arguments{
\item{strata}{the 'strata' dataset.}

\item{errors}{the 'errors' dataset containing the current precision constraints.}
}
\value{
the new 'errors' dataset containing the changed precision constraints
}
\description{
Function to propose a convenient set of precision constraints,
given the different variability of target variables Ys in the strata
}
\details{
The function requires in input the 'strata' dataset, plus an initial
set of precision constraints. It is suggested to define a 'neutral'
set, i.e. equal CVs for all variable, possibly differentiated
only with regard to the different domain levels
}
\examples{
data(beat.example)
errors[1,c(2:3)] <- c(0.02,0.02)
errors[2,c(2:3)] <- c(0.03,0.03)
errors
cv <- CVs_hint(strata,errors)
cv
}
