% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInputToAllocation_beat.1st.R
\name{prepareInputToAllocation_beat.1st}
\alias{prepareInputToAllocation_beat.1st}
\title{The function returns a dataframe, starting from the sampling frame (either universe or sample of a previous survey) with strata information.}
\usage{
prepareInputToAllocation_beat.1st(
  samp_frame,
  ID,
  stratum,
  dom,
  target,
  samp_weight = NULL
)
}
\arguments{
\item{samp_frame:}{A dataframe with the sampling frame (either universe or sample of another survey).}

\item{ID:}{name of the variable which is the identifier of the units in the sampling frame.}

\item{stratum:}{either name of the variable in samp_frame which is taken as the stratum, or the name of the variables which have to be concatenated to obtain the stratum. In this last case, the variables used to build the stratum are retained..}

\item{dom:}{name of the variable(s) in samp_frame which are the domain(s) of the estimates.}

\item{target:}{variable(s) in samp_frame which is the target of the estimate.}

\item{samp_weight:}{name of the variable in the sample of another survey specifying the sampling weights. Default is NULL, meaning that population is taken as the reference.}
}
\value{
A dataframe with strata information. The output dataframe can be used as input dataframe stratif for R2BEAT one-stage sample design (beat.1st)
}
\description{
The function returns a dataframe, starting from the sampling frame (either universe or sample of a previous survey) with strata information.
}
\examples{
# ---------------- 
# from population
# ---------------- 
data("pop")
samp_frame=pop
ID="id_ind"
stratum="stratum"
dom=c("region", "province" )
target=c("unemployed")
a <- prepareInputToAllocation_beat.1st(samp_frame = samp_frame, 
                                       ID=ID, 
                                       stratum=stratum, 
                                       dom=dom,
                                       target=target)

# ---------------- 
# from sample
# ----------------
data("sample")
samp_frame=samp
ID="id_ind"
stratum="stratum"
dom=c("region", "province" )
target=c("unemployed")
samp_weight="weight"
b <- prepareInputToAllocation_beat.1st(samp_frame = samp_frame, 
                                       ID=ID, 
                                       stratum=stratum, 
                                       dom=dom,
                                       target=target, 
                                       samp_weight = samp_weight)
}
