\name{sens_names}
\alias{sens_names}
\title{
Function for better presentation of sensitivity information
}
\description{
This function allows to obtain a better presentation of sensitivity information, i.e. with the names of the variables and of the domains

}
\usage{
sens_names(s,target_vars,strata)
}

\arguments{
 \item{s}{The sensitivity output of the allocation step.}
 \item{target_vars}{target variables.}
 \item{strata}{strata dataframe used in the allocation step..}
}

\value{
A dataframe containing Sensitivity information
}


\author{Giulio Barcaroli}
